+++ .travis.yml
@@ -2,4 +2,4 @@ language: node_js
 node_js: node
 
 after_success: 
-- echo "Done"
+- npm run coverage
