+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "start": "node app.js",
     "test": "mocha",
-    "coverage": "coveralls"
+    "coveralls": "coveralls"
   },
   "repository": {
     "type": "git",
@@ -34,7 +34,7 @@
     "chai": "^4.2.0",
     "chai-http": "^4.2.0",
     "coveralls": "^3.0.2",
-    "eslint": "^5.8.0",
+    "eslint": "^5.9.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
     "mocha-lcov-reporter": "^1.3.0"
